refactor: extract component styles into dedicated files
- Move inline styles from all components to separate style files
- Create dedicated style files for each screen/component
- Rename colors.tsx to globalStyles.tsx for better clarity
- Add new style modules (aboutStyles, branchStyles, etc.)
- Implement consistent style organization across the application
- Reduce component file sizes for better readability

This change implements Issue #1: Refactor separate styles into their own files
