+++ app/(tabs)/about.tsx
@@ -1,21 +1,11 @@
-import { Text, View, StyleSheet } from 'react-native';
+import { Text, View } from 'react-native';
+import { aboutStyles } from '../styles/aboutStyles';
 
 export default function AboutScreen() {
   return (
-    <View style={styles.container}>
-      <Text style={styles.text}>Luxe Salon</Text>
+    <View style={aboutStyles.container}>
+      <Text style={aboutStyles.text}>Luxe Salon</Text>
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#25292e',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    color: '#fff',
-  },
-});
