+++ app/_layout.tsx
@@ -5,14 +5,13 @@ import {
   TouchableOpacity,
   View,
   Text,
-  StyleSheet,
   Modal,
   FlatList,
   TouchableWithoutFeedback,
 } from 'react-native';
 import { useRouter, usePathname } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
-
+import { navigationStyles as styles } from './styles/navigationStyles';
 // Define valid route paths
 type AppRoute =
   | '/'
@@ -312,87 +311,3 @@ export default function RootLayout() {
   );
 }
 
-const styles = StyleSheet.create({
-  headerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    backgroundColor: 'white',
-    height: 60,
-    paddingHorizontal: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: '#e0e0e0',
-  },
-  homeButton: {
-    padding: 8,
-    marginRight: 8,
-  },
-  breadcrumbContainer: {
-    flex: 1,
-  },
-  breadcrumb: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  breadcrumbText: {
-    fontSize: 16,
-    color: '#6A1B9A',
-  },
-  pageTitle: {
-    fontSize: 18,
-    color: '#333',
-    fontWeight: '600',
-  },
-  menuButton: {
-    padding: 8,
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'flex-start',
-  },
-  menuContainer: {
-    backgroundColor: 'white',
-    width: '80%',
-    maxWidth: 300,
-    height: '100%',
-    alignSelf: 'flex-end',
-    borderTopLeftRadius: 10,
-    borderBottomLeftRadius: 10,
-    elevation: 5,
-    shadowColor: '#000',
-    shadowOffset: { width: -2, height: 0 },
-    shadowOpacity: 0.25,
-    shadowRadius: 5,
-  },
-  menuHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: '#e0e0e0',
-  },
-  menuTitle: {
-    fontSize: 20,
-    fontWeight: '600',
-    color: '#333',
-  },
-  menuList: {
-    flex: 1,
-  },
-  menuItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: '#f0f0f0',
-  },
-  menuItemIcon: {
-    marginRight: 15,
-  },
-  menuItemText: {
-    fontSize: 16,
-    color: '#333',
-  },
-});
