+++ app/screens/booking-complete.tsx
@@ -1,34 +1,43 @@
-// screens/booking-complete.tsx
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { View } from 'react-native';
 import { Text, Button, Card, Avatar } from 'react-native-paper';
 import { useRouter } from 'expo-router';
+import { bookingCompleteStyles } from '../styles/bookingStyles';
 
 export default function BookingComplete() {
   const router = useRouter();
 
   return (
-    <View style={styles.container}>
-      <Avatar.Icon icon="check-circle" size={120} style={styles.successIcon} color="#fff" />
+    <View style={bookingCompleteStyles.container}>
+      <Avatar.Icon
+        icon="check-circle"
+        size={120}
+        style={bookingCompleteStyles.successIcon}
+        color="#fff"
+      />
 
-      <Card style={styles.card}>
+      <Card style={bookingCompleteStyles.card}>
         <Card.Content>
-          <Text variant="headlineMedium" style={styles.heading}>
+          <Text variant="headlineMedium" style={bookingCompleteStyles.heading}>
             Booking Confirmed!
           </Text>
 
-          <Text style={styles.message}>
+          <Text style={bookingCompleteStyles.message}>
             Your appointment has been successfully booked. We&apos;ve sent a confirmation email with
             all the details.
           </Text>
 
-          <Text style={styles.reminderText}>
+          <Text style={bookingCompleteStyles.reminderText}>
             Please arrive 10 minutes before your appointment time.
           </Text>
         </Card.Content>
       </Card>
 
-      <Button mode="contained" onPress={() => router.push('/(tabs)/booking')} style={styles.button}>
+      <Button
+        mode="contained"
+        onPress={() => router.push('/(tabs)/booking')}
+        style={bookingCompleteStyles.button}
+      >
         Book Another Appointment
       </Button>
       <Button
@@ -37,45 +46,10 @@ export default function BookingComplete() {
           // In a real app, this would navigate to a bookings management screen
           router.push('/screens/booking-management');
         }}
-        style={styles.button}
+        style={bookingCompleteStyles.button}
       >
         View My Bookings
       </Button>
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  successIcon: {
-    backgroundColor: '#4CAF50',
-    marginBottom: 24,
-  },
-  card: {
-    width: '100%',
-    marginBottom: 24,
-  },
-  heading: {
-    textAlign: 'center',
-    marginBottom: 16,
-    color: '#4CAF50',
-  },
-  message: {
-    textAlign: 'center',
-    fontSize: 16,
-    marginBottom: 16,
-  },
-  reminderText: {
-    textAlign: 'center',
-    fontStyle: 'italic',
-  },
-  button: {
-    marginVertical: 8,
-    width: '100%',
-  },
-});
