+++ app/screens/date-time-selection.tsx
@@ -1,11 +1,11 @@
 // screens/date-time-selection.tsx
 import React, { useState, useEffect } from 'react';
-import { StyleSheet, View, ScrollView } from 'react-native';
+import { View, ScrollView } from 'react-native';
 import { Text, Card, Button, Title, Chip, ActivityIndicator } from 'react-native-paper';
 import { Calendar } from 'react-native-calendars';
 import { useRouter, useLocalSearchParams } from 'expo-router';
 import { CommandResult } from '../utils/voice-command-parser';
-
+import { dateTimeSelectionStyles as styles } from '../styles/dateTimeSelectionStyles';
 // Types
 type TimeSlot = {
   id: string;
@@ -208,59 +208,3 @@ export default function DateTimeSelection() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16,
-  },
-  heading: {
-    marginBottom: 16,
-    textAlign: 'center',
-  },
-  voiceCard: {
-    marginBottom: 16,
-    backgroundColor: '#e3f2fd',
-  },
-  voiceDetected: {
-    fontWeight: 'bold',
-    marginBottom: 4,
-  },
-  dateTitle: {
-    marginTop: 16,
-    marginBottom: 8,
-  },
-  timeTitle: {
-    marginTop: 24,
-    marginBottom: 16,
-  },
-  timeSlotsContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    marginBottom: 24,
-  },
-  timeChip: {
-    margin: 4,
-    backgroundColor: '#ffffff',
-  },
-  selectedTimeChip: {
-    backgroundColor: '#2196F3',
-  },
-  selectedTimeText: {
-    color: '#ffffff',
-  },
-  unavailableTimeChip: {
-    backgroundColor: '#f0f0f0',
-  },
-  loader: {
-    marginVertical: 24,
-  },
-  noTimesMessage: {
-    textAlign: 'center',
-    marginVertical: 24,
-    fontStyle: 'italic',
-  },
-  continueButton: {
-    marginVertical: 24,
-    paddingVertical: 8,
-  },
-});
