+++ app/styles/aboutStyles.tsx
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export const aboutStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#25292e',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    color: '#fff',
+  },
+});
