+++ app/styles/globalColors.tsx
