+++ app/styles/notFound.tsx
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+export const notFoundStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#25292e',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+
+  button: {
+    fontSize: 20,
+    textDecorationLine: 'underline',
+    color: '#fff',
+  },
+});
